updated index poster=
poster="images/swing-image.jpg" to poster="images/swing-image.png" 

To bridge the gap in the uploading of the video I originally used VLC to screenshot the video image but the differences fromposter=image to video was to great so i used "any DVD" to get the closest screenshot to the beginning of the video. Resulting in a png file.